[IMP] ..multi_client: methods and fields in mixin
