+++ microsoft_outlook_multi_client/models/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
+from . import microsoft_outlook_mixin
 from . import ir_mail_server
